Modify call_hash to use obj ids and mode bits; Cache inspect frame
The pretty print of the fields and constraints to use as a
key in the randomization cache is expensive. Since each call
has unique objects for the fields and constraints we can use
those instead. This also grabs the obj id of the with constraint
(block?) and the expressions/constraints within.

Saving the inspect.stack() frame in the randobj_interposer causes
deepcopy/pickle issues in the Randomization call. I'm not entirely
sure why yet, but saving the filename str and lineno int in the
SourceInfo and saving that to reuse later instead seems safer and
avoids the fairly large penalty I saw in py-spy.