Reorder set_val calls in post_rand; Add non-rand values to call_hash
The segmented_randomization test was reusing constraints with
nested vsc.attr objects, which are non-rand and change the
constraints with each call if they are different. Added a
bit to the pretty printer to grab that.

The order of set_val in nested post_randomize calls was
setting values from the parent randomization call after
the post-rand randomization call. This worked before caching
b/c they were the same Boolector var, but now they are
distinct. Swapped the order in FieldCompositeModel since
it seemed to be backward.