Fix pre_rand issues and add unit test; Minor cleanup
I think it's possibly and intended that pre_randomize could
change the constraints for a randomize call, so it should
be called everytime before generating the call_hash.

This causes some issues with deepcopy since rand_obj
was trying to prevent deepcopying self in the parent
reference of other objects. Workaround for now is
just to skip deepcopying the Boolector var object in
FieldScalarModel. This might not be the correct thing
to do, but passes unit tests for now.