Skip nested soft constraints when building hard constraints
Now that nested soft constraints are recreated as individual, soft
constraints the hard constraint building needs to skip them and
soft constraints need to be built with the soft flag enabled.
This happens both in the rand_set_node_builder and back in
randomizer when it constructs the list of python and Boolector
constraint tuples. The randomizer call should be picking up
cached objects, but still needs the flag to not return None
for the object. Unless there's a different way to do so?