Remove high-priority, soft weight constraint from distributions
This removes the soft constraint created after building the
hard constraints for a distribution. It also reverses the
unittest meant for testing that distribution non-zero weights
can override soft constraints.

This saves off the weight list and a function to calculate them
into ConstraintDistScopeModel, in the hopes that at some point
recalling the randomization model and swizzler would re-randomize
the distribution weight targeted.

This also adds enough type hints to make it easier to track down
target_range and distribution weight list usage.