Make solve_order/toposort predictable
Using solve_order() on a list could cause randomization of list
members in an unpredictable order (see #221). This was due to
toposort using sets and unordering the dictionary passed to it.
This now appends the toposorted list fields in the same order
that the randset had originally.

Add testcase passing multiple lists in an ordered randset.