Merge dists when combining randsets
When two randsets with dist constraints combined the dists from the
current randset were getting lost. This meant that for two related
variables with dists that only the last dist was ever used in the
randomization/swizzler.

This adds or merges the dist_field_m dict entries to the existing
randset to retain them.

Added a test case that ties two variables together with an enable bit
conditional.