swizzler: add dist constraint retry and fallback
Extends the distribution constraint exploration of the swizzler
to retry the values it selects some number of times (4 for now)
and if those all fail, fall back to randomizing the field bits
as any other variable.

The swizzler was only picking one value, and if the given
constraints excluded most values then it resulted in the solver
tossing it and leaving the field unrandomized, which in many
cases lets the solver default to 0 or some other value repeatedly.

See #243